+++ src/main/java/de/samply/db/model/Inquiry.java
@@ -1,5 +1,7 @@
 package de.samply.db.model;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonInclude;
 import de.samply.converter.Format;
 import de.samply.teiler.TeilerConst;
 import jakarta.persistence.Column;
@@ -10,7 +12,7 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import java.time.Instant;
+import jakarta.persistence.Transient;
 import org.hibernate.annotations.Formula;
 
 @Entity
@@ -46,14 +48,22 @@ public class Inquiry {
   @Formula("TO_CHAR(archived_at, 'YYYY-MM-DD HH24:MI')")
   private String archivedAt;
 
-  @Formula("(SELECT e.template_id FROM samply.query q JOIN samply.query_execution e ON q.id = e.query_id ORDER BY e.executed_at DESC LIMIT 1)")
+  @Formula("(SELECT e.template_id FROM samply.query q JOIN samply.query_execution e ON q.id = e.query_id WHERE q.id = id ORDER BY e.executed_at DESC LIMIT 1)")
   private String templateId;
 
-  @Formula("(SELECT TO_CHAR(e.executed_at, 'YYYY-MM-DD HH24:MI') FROM samply.query q JOIN samply.query_execution e ON q.id = e.query_id ORDER BY e.executed_at DESC LIMIT 1)")
+  @Formula("(SELECT TO_CHAR(e.executed_at, 'YYYY-MM-DD HH24:MI') FROM samply.query q JOIN samply.query_execution e ON q.id = e.query_id WHERE q.id = id ORDER BY e.executed_at DESC LIMIT 1)")
   private String lastExecutedAt;
 
-  @Formula("(SELECT er.error FROM samply.query q JOIN samply.query_execution e ON q.id = e.query_id LEFT JOIN samply.query_execution_error er ON e.id = er.query_execution_id ORDER BY e.executed_at DESC LIMIT 1)")
+  @Formula("(SELECT er.error FROM samply.query q JOIN samply.query_execution e ON q.id = e.query_id LEFT JOIN samply.query_execution_error er ON e.id = er.query_execution_id WHERE q.id = id ORDER BY e.executed_at DESC LIMIT 1)")
   private String error;
+
+  @JsonInclude
+  @Transient
+  private String executionFileUrl;
+
+  @JsonIgnore
+  @Formula("(SELECT e.id FROM samply.query q JOIN samply.query_execution e ON q.id = e.query_id WHERE q.id = id ORDER BY e.executed_at DESC LIMIT 1)")
+  private Long queryExecutionId;
 
   public Long getId() {
     return id;
@@ -141,6 +151,22 @@ public class Inquiry {
 
   public void setError(String error) {
     this.error = error;
+  }
+
+  public String getExecutionFileUrl() {
+    return executionFileUrl;
+  }
+
+  public void setExecutionFileUrl(String executionFileUrl) {
+    this.executionFileUrl = executionFileUrl;
+  }
+
+  public Long getQueryExecutionId() {
+    return queryExecutionId;
+  }
+
+  public void setQueryExecutionId(Long queryExecutionId) {
+    this.queryExecutionId = queryExecutionId;
   }
 
 }